Merge duplicate elephant templates to a new inherited template. Patch by fatherbushido, refs #1487.
Prevent these elephants from attacking structures, which can happen incidentally if the animal is attacked by garrisoned arrows that were meant to hit attacking players and because that is unexpected to players and had been reported as a bug and cheating players.
Prevent them from attacking ships, because they can't reach them in most cases.

Unifty their Attack stats that had neglible differences, besides the missing Crush damage of the North African Elephant variant, forgotton to make consistent in rP9487 with rP8363.
Add the large passability class to the North African and Asian elephant, which was forgotton by rP16868.



git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@19295 3db68df2-c116-0410-a063-a993310a9797