Split the automatic aiming function of PerfomAttack for ranged attacks.
Summary:
PerformAttack uses an automatic aiming logic for ranged attacks.
Splitting it would allow to fix some issues with that function, to test it individually, make it usable as a globalscript for other logic (or even to disable the automatic aiming based on template).
That patch is just about the splitting, I didn't change it.
Also use only the max range and not the elevation adapted one to scale the spread (which makes more sense and also avoid computation).

Test Plan: send a javelin?

Reviewers: O2 JS Simulation, wraitii, elexis

Reviewed By: O2 JS Simulation, wraitii, elexis

Subscribers: Vulcan, elexis, leper

Differential Revision: https://code.wildfiregames.com/D272

git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@19366 3db68df2-c116-0410-a063-a993310a9797