Roman Entrenched Army Camp balancing.
Prevent building it in player owned territory, so as to
  highlight its usage as a building to dominate the enemy in his own territory and
  prevent the player from building it as a defensive structure right besides a fortress and civic center.

Increase CapturePoints from 500 to 1500, and RegenRate to 10, so that it doesn't convert to gaia in 14 seconds but 50.
Reduce GarrisonRegenRate from 5 to 3.
Add minimum distance between two camps.
Increase cost to 500 wood and buff crush armor from 10% to 19%.
Reduce maximum range from 80 meters to 50 meters.

Differential Revision: https://code.wildfiregames.com/D405
Reviewed By: fatherbushido
Reported by PhyZic, discussed with Grugnas.



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@19505 3db68df2-c116-0410-a063-a993310a9797