Prohibit cheats if cheats are disabled, i.e. developer cheats to control enemy units, revealing the map and promoting units too.
Open the developer overlay only if cheats are enabled (which is always the case in singleplayer mode and only the case in multiplayer mode if explicitly enabled).
(This doesn't make it harder for developers to "debug", since they previously had to remember to disable the rated game setting too. Also every bug had to be reproduced with the replay either way)

Reverts rP17288 which reverted rP17282 (since as of rP19557, the autostart gamesetup enables cheats too).

Differential Revision: https://code.wildfiregames.com/D455
Fixes #3551
Reviewed By: echotangoecho
Agreed with Itms and Imarok in the last staff meeting.



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@19558 3db68df2-c116-0410-a063-a993310a9797