Manual update of the translation templates, in order to complete r19669.
This update doesn't ignore the line changes in .pot files (so it's more noisy but it can be useful for translators, who will now have the correct file numbers in the translation context) and I used the opportunity to update the translations now that we are preparing for a freeze.

The translation lint will be enabled after those templates go through Transifex and the dennis-ignore comment propagates.


git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@19670 3db68df2-c116-0410-a063-a993310a9797