Modern style cleanup.
Unify three textPanel style copies in a new ModernTextPanel theme,
following the Credits page in rP17830 and the Splash Screen in rP12735 and the Manual in rP7714.

Remove the unused textcolor_selected and the (because of a source/gui bug) defacto unused font property.
Rudimentary scrolling support for the Splash Screen.
Remove now empty files for the time being.



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@19992 3db68df2-c116-0410-a063-a993310a9797