Remove unneeded GetMucMessageCount from the XmppClient.
The affected function was imprecise, because it called a playerlist update whenever a "chat" level message was received instead
of only updating the playerlist if it's actually influencing the displayed playerlist.
When there is a chat message, lobby subject change or user-role change, there is no need to update the list.

Differential Revision: https://code.wildfiregames.com/D671
Refs #3386, rP16997
Reviewed By: fpre / ffffffff



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@20040 3db68df2-c116-0410-a063-a993310a9797