Move placeable templates out of special/.
special/ should not contain placeable templates. Now we can switch to a blacklist
approach for placeable templates in Atlas. This should make it easier for modders
to use a different folder structure than the public mod and having their templates
show up in Atlas without having to modifiy templates.
Non-placeable templates should exist in only a few different locations which are
unlikely to change often, and are relatively unlikely to be added regularly by
modders.

Reviewed By: elexis; some disagreement with FeXoR and Stan on the naming of the trigger points
Differential Revision: https://code.wildfiregames.com/D877

git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@20181 3db68df2-c116-0410-a063-a993310a9797