Switch back to an unplaceable filter for templates.
This removes FindAllPlaceableTemplates, replaces the few uses of it by FindAllTemplates,
and makes that ignore all templates starting with special/ in addition to those starting
with template_.

Now modders can use entirely different template organization schemes (more folders, different
folders, etc) without having to edit a file that was never well documented.

In conjunction with a few of the template moving patches preceding this rubble/ and other/catafalque
are now placeable. The former now does not decay anymore and users that want that should use the
decay| filter, the latter will be taken care of in #4762.

Return to making FindAllTemplates return all placeable templates again (switch to unplaceable filter).

To reiterate the main point: Only templates starting with special/ or template_ will not show up as
placeable in Atlas (or show up to code querying for all (placeable) templates. If you want to add more
of those use one of these naming schemes (and possibly subfolders in special/).

Reviewed By: fatherbushido
Differential Revision: https://code.wildfiregames.com/D935

git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@20246 3db68df2-c116-0410-a063-a993310a9797