Unify duplicate Corsica & Sardinia island generation, refs rP11266.
The code for the other island was copied in a varying order with only the values changed to 1 - x.
Unite the island, subisland, creeks, beaches, main relief, first plateau, second plateau and passagess code under a single loop to display the logical relation
instead of scattering many paragraphs and executing unrelated (starting base) code in between.

Cleanup affected createArea calls as in rP20301.



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@20306 3db68df2-c116-0410-a063-a993310a9797