Remove SemiRandomElevationPainter from Pyrenean Sierra, which is a painful straight copy of the SmoothElevationPainter introduced in rP12248, refs #4805.
It should have been added to the painters file instead of to the map if a copy was acceptable.
Add the only simple capability it added (an unnoticeable random elevation difference to each vertex) to the SmoothElevationPainter to make it universally accessible.
That painter also had a bug, because it replaced the roughness with 5 if 0 was given.



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@20354 3db68df2-c116-0410-a063-a993310a9797