Migrate to premake5, fixes #3729. Make pyrogenesis and tests depend on Collada, fixes #404.
This commit includes custom modules for CxxTest and pkg-config support that can be improved upon in the future. It also includes all necessary changes to the build scripts, but the new premake5 features are not fully exploited yet.

With this change, premake5 becomes the default, but CI scripts on Unix will continue using premake4 for a while, in order to avoid regressions.

Includes code by zsol.
Reviewed by: wraitii, leper
Differential Revision: https://code.wildfiregames.com/D72


git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@20381 3db68df2-c116-0410-a063-a993310a9797