Extend the cubicInterpolation function to consume a tension argument allowing to modulate the smoothness of the interpolation.
Thereby unify the chordal Catmull-Rom spline interpolation of the ClumpPlacer (C++ rP2378, JS rP9096),
the copy of that in the PathPlacer (rP11152, refs #892) and
the centripetal Catmull-Rom spline of the bicubicInterpolation function from rP18925, refs #4218
and don't claim the latter to be a uniform Catmull-Rom spline.

Reviewed in part by fatherbushido, discussed in rP18925.



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@20383 3db68df2-c116-0410-a063-a993310a9797