rmgen cleanup.
Remove placeTerrain, and createObjectGroup from Map, because that class should only have getters and setters, refs #4804.
Unify it with the global functions of the same name.
Let Map.createArea only construct and register the Area object with the Map, but do the place and painter logic in the global function.



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@20387 3db68df2-c116-0410-a063-a993310a9797