Introduce a new hotkey to order only one unit from the current selection.
This new hotkey, by default assigned to alt, makes it so that only one unit receives the order.
This can be used to easily order a group of units to build different buildings at the same time, or attack different enemies.
It's intended to make the usual "select a unit, order, select another unit, order" micro much faster.


Reviewed By: Imarok

Differential Revision: https://code.wildfiregames.com/D308


git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@20441 3db68df2-c116-0410-a063-a993310a9797