Fix some unintended obliteration in rP20420.
In the Edge Seas variant, fix a wrong horizontal/vertical negation and the direction of the riversize randomization.
Fix the wrong river location translation in that variant in r20429, refs #4851.
Prevent resource collisions on Unknown maps by adding back the large territory marking (after the small territory marking prior to the terrain generation).
Slightly improve river painting performance by not generating the meandering of the second shoreline if that is outside of the map boundary.



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@20465 3db68df2-c116-0410-a063-a993310a9797