Split uncommon random map playerbase terrain code from the common playerbase starting entities code. This allows replacing the common code with a library call in only one or two hunks each.
On Archipelago move the random resources closer to the playerbase to make the map more enjoyable.
Use default/skirmish templates for the docks on Islands/Migration and the market on Polar Sea while at it.
Abort map generation if the dock placement failed.
Remove clForest dummy variable on Polar Sea following the removal of references to map defined tileclasses in the library following rP20405.



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@20476 3db68df2-c116-0410-a063-a993310a9797