Change gathering rally points to use gather nstead of gather-near-position when possible.
This makes rallypoint flags nicer and is generally more efficient for unitAI too.
The behaviour falls back to gathering near position for any entity with a "speed" component, which is set by GuiInterface depending on having UnitMotion. This thus assumes entities without UnitMotion do not move. However, Gather falls back to gather near position in case of failure, so it ought to be fine in general.

Reviewed By: wraitii

Patch By: temple

Differential Revision: https://code.wildfiregames.com/D1014


git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@20571 3db68df2-c116-0410-a063-a993310a9797