Delete Savegame format version, refs rP10454.
It is unneeded because we won't add backwards compatibility for previous formats and
because the engine version check from rP16906 is sufficient and more accurate in testing compatibility of savegames of different releases.
The simpler code allows unifying the savegame version check with the check used by the replay menu (to be used in multiplayer gamesetups).

Differential Revision: https://code.wildfiregames.com/D1131
Reviewed By: wraitii
Agreed with: Imarok
Discussed with: Philip



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@20729 3db68df2-c116-0410-a063-a993310a9797