Merge InitMap, initTerrain and resetTerrain successors from rP20894.
Lets the random map script decide which baseTerrain and baseHeight it wants to use before initializing the map,
rather than painting what is in the JSON file, then changing ones mind and painting something else on top of it.
This cleaner heightmap initialization incidentally addresses the uninitialized border of the heightmap in that commit (excluding the Unknown).



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@20900 3db68df2-c116-0410-a063-a993310a9797