Fix wrong PathPlacer failfraction computation in rP11152, refs #892.
It unintentionally used the dist variable of the inner loop, not the dist variable reflecting the path length.
It was never noticeable since the few PathPlacers called with Constraints either had failCount 0 or 
exceeded the allowed failures by a factor of thousands.
All PathPlacer calls were intended to succeed or fail completely.
Hence set the default failFraction to 0.



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@20969 3db68df2-c116-0410-a063-a993310a9797