Implement ElevationBlendingPainter which interpolates the height of the given area with the desired height.
Primary use case, as represented on Caledonian Meadows, is creating a path through impassable area, mountains or water, refs #4952.
Supersedes placeRandomPathToHeight from the heightmap library, refs #3764.
The painter has the advantage that it can be applied to arbitrary areas with arbitrary forms of smoothing (or no smoothing).
Replace the rectangularSmoothToHeight calls that only smooth an area with a SmoothElevationPainter call.
Replace placeRandomPathToHeight which is a copy of the the RandomPathPlacer from rP20879, refs #4805.



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@21182 3db68df2-c116-0410-a063-a993310a9797