Cleanup random map Area prototype following rP9271.
Delete areaID and just identify the Area object by the actual JS object reference.
Delete all references from Areas in the RandomMap object,
especially the area property of the RandomMap object became invalid after creating a new area.
This was replaced with a cache in the Area object and a contains function.
Add getPoints getter, so that the Area property is not accessed directly.
Deepfreeze points to prevent accidental overwrites.
Implement StayAreasConstraint and extend currently unused AvoidAreaConstraint to accept multiple areas.



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@21296 3db68df2-c116-0410-a063-a993310a9797