Impose a population limit for gaia units on Jebel Barkal to prevent lag of doom.
Have strongly garrisoned buildings in the city capturable to incentivize players to break the walls.

Use "probabilistic rounding" for unit counts, so that on a 1v1 with a normal mapsize on medium difficulty,
don't round to 0 elephants per stable in the first wave as it is too few and
don't round to 1, because that were too many.

Place Sentry towers in easy difficulty instead of none.



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@21735 3db68df2-c116-0410-a063-a993310a9797