In preparation of renaming and grouping main.cpp shutdown variables:
Declare the g_GameRestarted hack from rP4520 in main.cpp, because it is main.cpp and Game.cpp who provide it's value.
Move the comment so that the connotation becomes clear when reading main.cpp.

Remove the extern declaration introduced to the Renderer in rP12802 that became obsolete with rP12817.



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@21816 3db68df2-c116-0410-a063-a993310a9797