Main.cpp cleanup.
Use an enum to indicate the type of engine shutdown instead of three bools.
State in the comments that the program is restarted within the same process.

In preparation of introducing an IsQuitRequested function (which shall not be named is_quit_requested as stressed by Vladislav):
Rename kill_mainloop to QuitEngine, restart_mainloop to RestartEngine, restart_mainloop_in_atlas to StartAtlas to not break consistency.
Don't rename RestartInAtlas in JS just now.
Group declarations at the top of main.cpp.



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@21817 3db68df2-c116-0410-a063-a993310a9797