Remove 256 character limit from escapeText.
escapeText was added for the gamesetup and session chat in rP8494,
but the character limit is counterproductive for chatting with more than 2 sentences per line, when displaying a file path or other arbitrary escaped text,
not really an effective measure against lobby spam, nor was that option considered in any of the later calls except rP18534 and
thus more likely to cause unintended than intended code behavior and
should instead be implemented with max_length for input fields and linebreaks and smaller GUI object sizes for displaying, refs #5266.

Remove early returns that are redundant code and require more performance in the average case (non-empty string).

From Differential-Revision: https://code.wildfiregames.com/D720



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@21853 3db68df2-c116-0410-a063-a993310a9797