Inform users that they are subject to the mod.io Terms and Conditions and Privacy Policy and provide links following rP21759 / D1029. Refs GDPR #5257. Add link to feedback.wildfiregames.com, which demonstrates users how their data is going to be used.
Add link buttons to the Terms dialog (refs #4583), so that users are informed of the terms prior to becoming subject to them.
Don't link the DMCA / copyright terms of mod.io now since I found no UI with support for 3 buttons that doesn't appear unsatisfyingly ugly.
Adds 3 button captions and 1 tooltip string.

Differential Revision: https://code.wildfiregames.com/D1627
A Comment and a relatable patch in D1601 by smiley / (-_-).



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@21890 3db68df2-c116-0410-a063-a993310a9797