Clean up and speed up the water manager distance computations
The water manager computations for "fancy water effects" have always been quite slow. I've updated one of the functions to be much faster, and the other doesn't need to be called (apparently, since I removed coastal foam, which tbh I don't remember doing).

This should all be redone entirely to be honest, as it's generally terrible, but in the short-term™ this makes this function almost usable in real-time.

Differential Revision: https://code.wildfiregames.com/D78


git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@22006 3db68df2-c116-0410-a063-a993310a9797