Always leave/reenter the top-most FSM state when SetNextState is called with the current FSM state.
When SetNextState is called with the current state, it will leave and re-enter so that any relevant cleanup will get called and things work as expected.
It will prevent possibly awkward code issues in UnitAI, which already called SetNextStateAlwaysEntering in a number of places to work around them.

This should most likely have been the default behaviour in the first place.

Reviewed By: Itms

Differential Revision: https://code.wildfiregames.com/D1488


git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@22023 3db68df2-c116-0410-a063-a993310a9797