Fix some animals "roaming" state not working correctly (units seem stuck in place)
Due to an issue in rP22352, animals incorrectly tried going towards the roaming point instead of away from it.

With that fixed, MovementUpdate still did not trigger as the min and max range were the same.
Use -1 as max range (= infinite) instead since we want to move arbitrarily away.

Having an infinite max range was broken in rP22345, this re-implements that.

Further, other calls with equal min and max range have been changed likewise.

This does not entirely fix whales, which run in other problems because of their large roaming range.

Differential Revision: https://code.wildfiregames.com/D1980


git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@22413 3db68df2-c116-0410-a063-a993310a9797