Fix max-range approach when treating a target as a circle.
As reported by @bb in #5512, catapults and other big-range units might not approach targets correctly when trying to get in range.
To be sure that we will be in range of a square obstruction approximated as a circle, we need to consider the inscribed circle and not the circumscribed circle.

Reported by: bb

Fixes #5512

Differential Revision: https://code.wildfiregames.com/D2086


git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@22495 3db68df2-c116-0410-a063-a993310a9797